+++ boot/authority-web-boot-starter/src/main/java/com/kewen/framework/boot/authority/biz/service/SysUserRoleService.java
@@ -3,7 +3,7 @@ package com.kewen.framework.boot.authority.biz.service;
 import java.util.List;
 import com.kewen.framework.boot.authority.biz.mapper.entity.SysUserRole;
 import com.baomidou.mybatisplus.extension.service.IService;
-import com.kewen.common.model.Role;
+import com.kewen.framework.base.common.model.Role;
 
 public interface SysUserRoleService extends IService<SysUserRole>{
 
@@ -18,5 +18,5 @@ public interface SysUserRoleService extends IService<SysUserRole>{
 
     int insertOrUpdateSelective(SysUserRole record);
 
-    List<Role> listUserRole(Integer userId);
+    List<Role> listUserRole(Long userId);
 }
